+++ examples/icd/_02_shin_algo.py
@@ -42,15 +42,17 @@ from mobgap.pipeline import GsIterator
 
 iterator = GsIterator()
 
-for gs_id, (gs, data), result in iterator.iterate(imu_data, reference_wbs):
+for (gs, data), result in iterator.iterate(imu_data, reference_wbs):
     icd = IcdShinImproved()
     icd.detect(data, sampling_rate_hz=sampling_rate_hz)
     result.ic_list = icd.ic_list_
     refined_gs_list, refined_ic_list = refine_gs(result.ic_list)
 
-    rgs_id, (_, refined_data), refined_result = iterator.with_subregion(refined_gs_list)
+    (refined_gs, refined_data), refined_result = iterator.with_subregion(refined_gs_list)
     refined_result.cad_per_sec = (
-        CadFromIc().calculate(refined_data, refined_ic_list.loc[rgs_id], sampling_rate_hz=sampling_rate_hz).cad_per_sec_
+        CadFromIc()
+        .calculate(refined_data, refined_ic_list.loc[refined_gs.id], sampling_rate_hz=sampling_rate_hz)
+        .cad_per_sec_
     )
 
     # for rgs_id, (refined_gs, refined_data), refined_result in iterator.iterate_subregions(refined_gs_list):