+++ tests/test_pipeline/test_gait_sequence_iterator.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from typing import Any
 
 import pandas as pd
 import pytest
@@ -6,6 +7,7 @@ from pandas.testing import assert_frame_equal
 
 from mobgap.pipeline import GsIterator, create_aggregate_df, iter_gs
 from mobgap.pipeline._gs_iterator import GaitSequence, WalkingBout
+from mobgap.pipeline._overwrite_typed_iterator import TypedIteratorResultTuple
 
 
 class TestGsIterationFunc:
@@ -60,15 +62,15 @@ class TestGsIterator:
 
         raw_results = iterator.raw_results_
         assert len(raw_results) == 2
-        assert raw_results[0].n_samples == 5
-        assert raw_results[0].s_id == "s1"
-        assert raw_results[1].n_samples == 5
-        assert raw_results[1].s_id == "s2"
+        assert raw_results[0].result.n_samples == 5
+        assert raw_results[0].result.s_id == "s1"
+        assert raw_results[1].result.n_samples == 5
+        assert raw_results[1].result.s_id == "s2"
 
-        assert iterator.n_samples_ == [5, 5]
-        assert iterator.s_id_ == ["s1", "s2"]
+        assert iterator.results_.n_samples == [5, 5]
+        assert iterator.results_.s_id == ["s1", "s2"]
 
-        inputs, input_dfs = zip(*iterator.inputs_)
+        inputs, input_dfs = zip(*(v.input for v in iterator.raw_results_))
         assert inputs == (("s1", 0, 5), ("s2", 5, 10))
         assert_frame_equal(input_dfs[0], pd.DataFrame({"data": [1, 2, 3, 4, 5]}, index=[0, 1, 2, 3, 4]))
         assert_frame_equal(
@@ -82,7 +84,9 @@ class TestGsIterator:
             n_samples: int
             s_id: str
 
-        aggregations = [("n_samples", lambda _, r: sum(r))]
+        aggregations = [
+            ("n_samples", lambda r: sum(v.result for v in GsIterator.filter_iterator_results(r, "n_samples")))
+        ]
 
         dummy_data = pd.DataFrame({"data": [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]})
         dummy_sections = pd.DataFrame({"start": [0, 5], "end": [5, 10], "gs_id": ["s1", "s2"]}).set_index("gs_id")
@@ -92,7 +96,7 @@ class TestGsIterator:
         for (_, d), r in iterator.iterate(dummy_data, dummy_sections):
             r.n_samples = len(d)
 
-        assert iterator.n_samples_ == 10
+        assert iterator.results_.n_samples == 10
 
     def test_default_agg_offsets(self):
         dummy_sections = pd.DataFrame({"start": [0, 5], "end": [5, 10], "wb_id": ["s1", "s2"]}).set_index("wb_id")
@@ -121,46 +125,63 @@ class TestAggregateDf:
     """Some isolated tests for the create_aggregate_df function."""
 
     @pytest.mark.parametrize("fix_cols", [[], ["start", "end"], ["start"]])
-    def test_basic_aggregation(self, fix_cols):
-        aggregate_df = create_aggregate_df(fix_gs_offset_cols=fix_cols)
+    @pytest.mark.parametrize("id_col_name", (None, "bla"))
+    def test_basic_aggregation(self, fix_cols, id_col_name):
+        aggregate_df = create_aggregate_df("test", fix_gs_offset_cols=fix_cols)
 
-        example_sequences = pd.DataFrame({"start": [0, 5], "end": [5, 10]}, index=["s1", "s2"]).rename_axis(
-            index="wb_id"
-        )
+        example_sequences = pd.DataFrame({"start": [0, 5], "end": [5, 10]}, index=["s1", "s2"]).rename_axis(index="id")
         example_inputs = [(s, pd.DataFrame()) for s in example_sequences.reset_index().itertuples(index=False)]
 
+        @dataclass
+        class DummyResultType:
+            test: Any
+
+        context = {"id_col_name": id_col_name} if id_col_name else {}
+
         example_results = [
-            pd.DataFrame({"start": [1, 3], "end": [3, 5]}),
-            pd.DataFrame({"start": [1, 3], "end": [3, 5]}),
+            TypedIteratorResultTuple(
+                "__main__", example_inputs[0], DummyResultType(pd.DataFrame({"start": [1, 3], "end": [3, 5]})), context
+            ),
+            TypedIteratorResultTuple(
+                "__main__", example_inputs[1], DummyResultType(pd.DataFrame({"start": [1, 3], "end": [3, 5]})), context
+            ),
         ]
 
-        aggregated = aggregate_df(example_inputs, example_results)
+        aggregated = aggregate_df(example_results)
 
         expected_start = [1, 3, 6, 8] if "start" in fix_cols else [1, 3, 1, 3]
         expected_end = [3, 5, 8, 10] if "end" in fix_cols else [3, 5, 3, 5]
+
+        real_id_col_name = id_col_name or "gs_id"
 
         assert_frame_equal(
             aggregated,
             pd.DataFrame(
                 {"start": expected_start, "end": expected_end},
-                index=pd.MultiIndex.from_product((example_sequences.index, [0, 1]), names=["wb_id", None]),
+                index=pd.MultiIndex.from_product((example_sequences.index, [0, 1]), names=[real_id_col_name, None]),
             ),
         )
 
     @pytest.mark.parametrize("fix_index_offset", [True, False])
     def test_index_offset(self, fix_index_offset):
-        example_sequences = pd.DataFrame({"start": [1, 5], "end": [5, 10]}, index=["s1", "s2"]).rename_axis(
-            index="wb_id"
-        )
+        example_sequences = pd.DataFrame({"start": [1, 5], "end": [5, 10]}, index=["s1", "s2"]).rename_axis(index="id")
         example_inputs = [(s, pd.DataFrame()) for s in example_sequences.reset_index().itertuples(index=False)]
 
+        @dataclass
+        class DummyResultType:
+            test: Any
+
         example_results = [
-            pd.DataFrame({"value": [1, 3]}, index=[1, 2]),
-            pd.DataFrame({"value": [2, 4]}, index=[3, 4]),
+            TypedIteratorResultTuple(
+                "__main__", example_inputs[0], DummyResultType(pd.DataFrame({"value": [1, 3]}, index=[1, 2])), {}
+            ),
+            TypedIteratorResultTuple(
+                "__main__", example_inputs[1], DummyResultType(pd.DataFrame({"value": [2, 4]}, index=[3, 4])), {}
+            ),
         ]
 
-        aggregate_df = create_aggregate_df(fix_gs_offset_index=fix_index_offset)
-        aggregated = aggregate_df(example_inputs, example_results)
+        aggregate_df = create_aggregate_df("test", fix_gs_offset_index=fix_index_offset)
+        aggregated = aggregate_df(example_results)
 
         if fix_index_offset:
             expected_index = [2, 3, 8, 9]
@@ -168,7 +189,7 @@ class TestAggregateDf:
             expected_index = [1, 2, 3, 4]
 
         expected_index = pd.MultiIndex.from_tuples(
-            zip(*(["s1", "s1", "s2", "s2"], expected_index)), names=["wb_id", None]
+            zip(*(["s1", "s1", "s2", "s2"], expected_index)), names=["gs_id", None]
         )
         assert_frame_equal(
             aggregated,
@@ -178,11 +199,5 @@ class TestAggregateDf:
             ),
         )
 
-    def test_wrong_datatype_input(self):
-        example_sequences = pd.DataFrame({"start": [0, 5], "end": [5, 10]}, index=["s1", "s2"]).rename_axis(
-            index="wb_id"
-        )
-        example_inputs = [(s, pd.DataFrame()) for s in example_sequences.reset_index().itertuples(index=False)]
-
-        with pytest.raises(TypeError):
-            create_aggregate_df()(example_inputs, "not a df")
+    # TODO: Add tests for subiteration
+    # TODO: Add tests for empty wbs.